fix: resolve TypeScript errors from structure refactoring
- notifications.ts: update severity-utils import to lib/utils/ path
- cleanup.ts: define MAX_BATCH_ITERATIONS locally in orphan cleanup method
- repositories/route.ts: coalesce nullable Zod fields with ?? undefined
- generate-openapi.ts: update import to lib/api/ path
- Remove stale .next type cache for old /api/image/ routes
